requirements.txt
zigzag-dse==3.6.0
torch
onnx
seaborn
matplotlib
